Changed defaults added script run times.
The defaults have been changed to only allow 1 user session at a time and 30 seconds between checks to keep the script more in line with the PMSF alt auth-bot. Current time is now scripted with each loop of the script, to display a print out. Minor formatting for better print out when using python 2.7.