Added config section for easier user setup/mods
Variables for the max number of sessions per user, time between database check cycles, and the database config have all been moved to a config file at the top of the script. Users no longer need to edit the script in multiple places to make changes, everything is now done in the config on the top of the script.